+++ src/main/java/by.epam.jwd.sak.avtobase/controller/impl/EditStatusRequestByDriver.java
@@ -19,7 +19,7 @@ public class EditStatusRequestByDriver implements Command {
     public void execute(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         if (req.getMethod().equals(POST)) {
             String status = "BEINPROGRESS";
-            Integer requestId = Integer.valueOf(req.getParameter(REQUEST_ID));
+            Long requestId = Long.valueOf(req.getParameter(REQUEST_ID));
 /*            Integer driverId = Integer.valueOf(req.getParameter(DRIVER));*/
             try {
 /*                factoryService.getDriversRequestsService().save(driverId, requestId);*/
