+++ src/main/java/by.epam.jwd.sak.avtobase/dao/CommentDao.java
@@ -11,7 +11,7 @@ public interface CommentDao {
 
     Comment save(Comment entity) throws DAOException;
 
-    List<Comment> findAllByUserId(Integer userId) throws DAOException;
+    List<Comment> findAllByUserId(Long userId) throws DAOException;
 
-    boolean delete(Integer id) throws DAOException;
+    boolean delete(Long id) throws DAOException;
 }
