+++ src/main/java/by.epam.jwd.sak.avtobase/service/RequestService.java
@@ -8,17 +8,17 @@ import java.util.Optional;
 
 public interface RequestService {
 
-    List<RequestDto> findAllRequestByUser(Integer userId) throws ServiceException;
+    List<RequestDto> findAllRequestByUser(Long userId) throws ServiceException;
 
     List<RequestDto> findAllRequest() throws ServiceException;
 
-    Integer create(RequestDto requestDto) throws ServiceException;
+    Long create(RequestDto requestDto) throws ServiceException;
 
-    Optional<RequestDto> findById(Integer id) throws ServiceException;
+    Optional<RequestDto> findById(Long id) throws ServiceException;
 
-    Integer update(RequestDto entity) throws ServiceException;
+    Long update(RequestDto entity) throws ServiceException;
 
-    boolean updateStatusById(Integer id, String status) throws ServiceException;
+    boolean updateStatusById(Long id, String status) throws ServiceException;
 
-    boolean delete(Integer id) throws ServiceException;
+    boolean delete(Long id) throws ServiceException;
 }
